Rename BrowserPanel -> TaskViewPanel
